Extract Merkle math from tlog_tiles into a new tlog_core crate
Slice 1 of #230: the c2sp.org/tlog-tiles spec covers an HTTP wire
format for tile-encoded transparency logs. Until now `tlog_tiles`
also carried the underlying RFC 6962 Merkle math (`Hash`,
`HashReader`, the proof builders/verifiers, the `Subtree` type +
the draft-ietf-plants-merkle-tree-certs subtree variants). That math
isn't part of `tlog-tiles` itself — it's the algorithm-only core
that every transparency-log spec in this workspace builds on.

This commit moves that nucleus into a new `tlog_core` crate and
migrates every internal consumer to import from it directly.

New crate `crates/tlog_core`:

- Carries `Hash`, `HASH_SIZE`, `EMPTY_HASH`, `HashReader`, `Proof`,
  `Subtree`, `LeafIndex`, `record_hash`, `node_hash`,
  `stored_hash_index` + variants, `tree_hash` + variants,
  `subtree_hash` + variants, `inclusion_proof` + verifiers,
  `consistency_proof` + verifiers, the subtree variants of all
  proofs, `evaluate_subtree_inclusion_proof`, and `TlogError` (now
  scoped to math errors only).
- Dependencies are `base64`, `serde`, `sha2`, `thiserror` only —
  no `ed25519-dalek`, `signed_note`, `length_prefixed`, `url`,
  `rand` dragged along.

Cleanups in `tlog_tiles`:

- New `CheckpointError` type local to `checkpoint.rs` covers
  `open_checkpoint` and `TreeWithTimestamp::sign` failure modes
  (`MissingVerifierSignature`, `InvalidTimestamp`, `OriginMismatch`,
  `Note(NoteError)`, `MalformedCheckpoint`, `Tlog(TlogError)`).
  Previously these were variants on `tlog_tiles::TlogError` even
  though they had nothing to do with the math layer.
- `Ed25519CheckpointSigner::new` is now infallible (`-> Self` instead
  of `-> Result<Self, TlogError>`); it never actually errored.
- `TlogTilesPendingLogEntry::ParseError` is now `std::io::Error`
  directly — it parses big-endian uint16 length-prefixed blobs and
  the IO error was the only failure path.

The `test_tree` integration test (which exercises the math against
tile-encoded storage) moved to `tlog_tiles/src/tile.rs` together with
the `TestTilesStorage` helper, since it depends on `Tile`,
`TileReader`, `TileHashReader`. Pure-math tests stayed in
`tlog_core`.

Migration: `tlog_witness`, `tlog_cosignature`, `static_ct_api`,
`bootstrap_mtc_api`, `bootstrap_mtc_worker`, `ct_worker`,
`generic_log_worker`, `witness_worker`, `tlog_tiles_wasm`, and
`integration_tests` all import moved symbols from `tlog_core`
directly. `tlog_witness` no longer depends on `tlog_tiles` at all.

The `fuzz` crate is unchanged: it already used deep-path imports
(`tlog_tiles::tile::Tile`, `tlog_tiles::checkpoint::CheckpointText`)
that are still valid.

Slices 2 (move `checkpoint.rs` to a new `tlog_checkpoint` crate)
and 3 (move `entries.rs` into `generic_log_worker`) are deferred
to follow-up PRs per the issue's staging plan.
