+++ crates/bootstrap_mtc_api/Cargo.toml
@@ -20,6 +20,7 @@ serde_with.workspace = true
 sha2.workspace = true
 signed_note.workspace = true
 thiserror.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 x509-cert.workspace = true
 x509_util.workspace = true
