+++ crates/bootstrap_mtc_worker/Cargo.toml
@@ -58,6 +58,7 @@ serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
 signed_note.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 worker.workspace = true
 x509-cert.workspace = true
