+++ crates/ct_worker/Cargo.toml
@@ -50,6 +50,7 @@ serde_json.workspace = true
 serde_with.workspace = true
 static_ct_api.workspace = true
 signed_note.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 worker.workspace = true
 x509-cert.workspace = true
