+++ crates/generic_log_worker/Cargo.toml
@@ -40,6 +40,7 @@ serde_json.workspace = true
 sha2.workspace = true
 signed_note.workspace = true
 thiserror.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 tokio.workspace = true
 worker.workspace = true
