+++ crates/static_ct_api/Cargo.toml
@@ -37,6 +37,7 @@ sha2.workspace = true
 signature.workspace = true
 signed_note.workspace = true
 thiserror.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 spki.workspace = true
 x509-cert.workspace = true
