+++ crates/tlog_cosignature/Cargo.toml
@@ -20,6 +20,7 @@ ed25519-dalek.workspace = true
 length_prefixed.workspace = true
 ml-dsa.workspace = true
 signed_note.workspace = true
+tlog_core.workspace = true
 tlog_tiles.workspace = true
 
 [dev-dependencies]
