+++ crates/tlog_witness/Cargo.toml
@@ -18,4 +18,4 @@ crate-type = ["rlib"]
 base64.workspace = true
 signed_note.workspace = true
 thiserror.workspace = true
-tlog_tiles.workspace = true
+tlog_core.workspace = true
