+++ crates/witness_worker/Cargo.toml
@@ -40,6 +40,7 @@ serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
 signed_note.workspace = true
+tlog_core.workspace = true
 tlog_cosignature.workspace = true
 tlog_tiles.workspace = true
 tlog_witness.workspace = true
